Changed the path dependence to serve-sofa-href
The path for the "serve-soft-hrtg" was problematic for me when using "npm install ambisonics". Github repos can be used directly using the username and repo, hence "Ircam-Rnd/serveSofaHrir" (but you might know that). The previous way led to npm ERR for me, maybe its just the network I'm on, but with the change on my fork, I was able to install the package using "npm install renaudfv/JSAmbisonics". I can send you the error log or the saved session from the Terminal if needed.