Changed main file in package.json
The main file config pointed to the bundle UMD file which was meant to be used as a global from HTML. 
This does not comply to the way node modules are typically used, especially when serve from NPM. Having the UMD file as an entry point leaded to not be able to call "require('ambisonics')" and expose to pre-browserified files. 
By using the dist transpired file, developper can access the ambisonics classes/modules and bundle as they want, in a more common way.